feat: add secret to the extended exec package
It's needed to report an unhandled rejection that is async takes place after
the whole stream got finished. In that case it's reported as successful but
the exception comes after stream got finished. As the execution state helper
already closed the report intervall of the status we need to do the http request
from within the engine. Therefore we need the secret to call the report endpoint.