feat: add new field 'additionConnectorRoot' to StreamSingleNodeResult
If the node is part of a node chain that was triggered by an additional connector,
this field will hold the UUID of the root node (the one with the additional
connector). This allows the frontend to differentiate between logs of normal
nodes and logs of nodes behind an additional connector.